Add missing changes, fix unix compilation warnings

+++ lib/http_perform.c
@@ -223,7 +223,7 @@ sf_bool STDCALL http_perform(CURL *curl,
             curl_easy_setopt(curl, CURLOPT_DEBUGDATA, &config);
 
             /* the DEBUGFUNCTION has no effect until we enable VERBOSE */
-            curl_easy_setopt(curl, CURLOPT_VERBOSE, 1);
+            curl_easy_setopt(curl, CURLOPT_VERBOSE, 1L);
         }
 
         if (header) {
@@ -243,7 +243,7 @@ sf_bool STDCALL http_perform(CURL *curl,
 
         // Post type stuffs
         if (request_type == POST_REQUEST_TYPE) {
-            res = curl_easy_setopt(curl, CURLOPT_POST, 1);
+            res = curl_easy_setopt(curl, CURLOPT_POST, 1L);
             if (res != CURLE_OK) {
                 log_error("Failed to set post [%s]", curl_easy_strerror(res));
                 break;
@@ -356,7 +356,7 @@ sf_bool STDCALL http_perform(CURL *curl,
             }
         }
 
-        res = curl_easy_setopt(curl, CURLOPT_SSLVERSION, SSL_VERSION);
+        res = curl_easy_setopt(curl, CURLOPT_SSLVERSION, (long)SSL_VERSION);
         if (res != CURLE_OK) {
             log_error("Unable to set SSL Version [%s]",
                       curl_easy_strerror(res));
@@ -370,14 +370,14 @@ sf_bool STDCALL http_perform(CURL *curl,
         } else {
             ocsp_check = SF_OCSP_CHECK;
         }
-        res = curl_easy_setopt(curl, CURLOPT_SSL_SF_OCSP_CHECK, ocsp_check);
+        res = curl_easy_setopt(curl, CURLOPT_SSL_SF_OCSP_CHECK, (long)ocsp_check);
         if (res != CURLE_OK) {
             log_error("Unable to set OCSP check enable/disable [%s]",
                       curl_easy_strerror(res));
             break;
         }
 
-        res = curl_easy_setopt(curl, CURLOPT_SSL_SF_OCSP_FAIL_OPEN, fail_open);
+        res = curl_easy_setopt(curl, CURLOPT_SSL_SF_OCSP_FAIL_OPEN, (long)fail_open);
         if (res != CURLE_OK) {
             log_error("Unable to set OCSP FAIL_OPEN [%s]",
                       curl_easy_strerror(res));
